LayoutTests/fullscreen/full-screen-document-background-color-expected.txt


RUN(frame.src = "resources/green.html")
EVENT(load)
EXPECTED (internals.documentBackgroundColor() == 'rgb(255, 255, 255)') OK
RUN(enterFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(0, 0, 0)') OK
RUN(exitFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(255, 255, 255)') OK

RUN(frame.src = "resources/empty.html")
EVENT(load)
RUN(enterFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(0, 0, 0)') OK
RUN(exitFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(255, 255, 255)') OK

RUN(frame.src = "resources/backdrop-green.html")
EVENT(load)
RUN(enterFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(0, 128, 0)') OK
RUN(exitFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(255, 255, 255)') OK

RUN(frame.src = "resources/backdrop-red-on-green.html")
EVENT(load)
RUN(enterFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(0, 128, 0)') OK
RUN(exitFullscreen())
EXPECTED (internals.documentBackgroundColor() == 'rgb(255, 255, 255)') OK
END OF TEST
